backend/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
psycopg2>=2.9.9
pytest-asyncio>=0.23.2
pytest-django>=4.7.0
dnspython
pymongo
django-multiselectfield
djangorestframework
SpeechRecognition
mongoengine
mongomock
gunicorn
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-dateutil
twilio
pydub
celery
django-celery-beat
redis
openpyxl
pytest-cov
black
isort
